backend/requirements.txt
certifi==2019.6.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
coverage==4.5.3
defusedxml==0.6.0
Django==2.2.3
django-allauth==0.39.1
django-cors-headers==3.0.2
django-extensions==2.1.9
django-oauth-toolkit==1.2.0
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-rest-swagger==2.0.7
djangorestframework==3.9.4
idna==2.8
itypes==1.1.0
Jinja2==2.10.1
MarkupSafe==1.1.1
oauthlib==3.0.1
openapi-codec==1.3.2
Pillow==6.1.0
psycopg2==2.8.3
python3-openid==3.1.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
simplejson==3.16.0
six==1.9.0
sqlparse==0.3.0
uritemplate==3.0.0
urllib3==1.25.3
